chore: Make InputReading variant like OutputCmd
Use the same enum strategy we use in OutputCmd for
InputReading data types.

Fixes #751

+++ crates/buttplug_client/src/device.rs
@@ -236,7 +236,7 @@ impl ButtplugClientDevice {
       .any(|x| *x.feature().feature_type() == FeatureType::RSSI)
   }
 
-  pub fn rssi_level(&self) -> ButtplugClientResultFuture<u32> {
+  pub fn rssi_level(&self) -> ButtplugClientResultFuture<i8> {
     if let Some(rssi) = self
       .device_features
       .iter()
